Refactor quiz logic to make the app more restful.
The app was modifying data in the server with each GET request, but this was an unintend
effect given a misuse of the get_context_data() of the QuestionView. The refactor aimed to
make every GET request in the core game logic idempotent, to move data manipulation logic to
the models, and to simplify the logic of the views removing as many data from the session
object as possible.