Change the way the results of quiz are stored on server.
Given the cleaner handle of data in the game logic now, it was possible to remove
a model entirely (GameSession), now all that logic is managed by the Answer model,
stored only when a user is logged in.

Anonymous users are able to play an compare it's results with the registered users,
but it's data won't persist.